优化clearBoard方法，释放文件

+++ server/src/main/java/io/github/springstudent/dekstop/server/clipboard/service/ClipboardServiceImpl.java
@@ -1,12 +1,15 @@
 package io.github.springstudent.dekstop.server.clipboard.service;
 
 import com.gysoft.jdbc.bean.Criteria;
+import io.github.springstudent.dekstop.common.utils.EmptyUtils;
 import io.github.springstudent.dekstop.server.clipboard.dao.ClipboardDao;
 import io.github.springstudent.dekstop.server.clipboard.pojo.Clipboard;
+import io.github.springstudent.dekstop.server.file.service.FileService;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * @author ZhouNing
@@ -18,9 +21,16 @@ public class ClipboardServiceImpl implements ClipboardService {
     @Resource
     private ClipboardDao clipboardDao;
 
+    @Resource
+    private FileService fileService;
+
     @Override
     public void clear(String deviceCode) throws Exception {
-        clipboardDao.deleteWithCriteria(new Criteria().where(Clipboard::getDeviceCode, deviceCode));
+        List<Clipboard> clipboards = clipboardDao.queryWithCriteria(new Criteria().where(Clipboard::getDeviceCode, deviceCode));
+        if (EmptyUtils.isNotEmpty(clipboards)) {
+            clipboardDao.deleteWithCriteria(new Criteria().where(Clipboard::getDeviceCode, deviceCode));
+            fileService.deleteFile(clipboards.stream().filter(clipboard -> EmptyUtils.isNotEmpty(clipboard.getFileInfoId())).map(Clipboard::getFileInfoId).collect(Collectors.toList()));
+        }
     }
 
     @Override
